requirements.txt
paho-mqtt>=1.6.1
flask>=2.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
aiomqtt>=1.0.0